# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: devrel@box.com
    name: "Box, Inc"
    url: https://box.dev
  description: "[Box Platform](https://box.dev) provides functionality to provide\
    \ access to content stored within [Box](https://box.com). It provides endpoints\
    \ for basic manipulation of files and folders, management of users within an enterprise,\
    \ as well as more complex topics such as legal holds and retention policies."
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  termsOfService: https://cloud.app.box.com/s/rmwxu64h1ipr41u49w3bbuvbsa29wku9
  title: Box Platform API
  version: 2.0.0
servers:
- url: https://api.box.com/2.0
paths:
  /folders/{folder_id}/items:
    get:
      description: |-
        Retrieves a page of items in a folder. These items can be files,
        folders, and web links.

        To request more information about the folder itself, like its size,
        please use the [Get a folder](#get-folders-id) endpoint instead.
      operationId: get_folders_id_items
      parameters:
      - description: |-
          The unique identifier that represent a folder.

          The ID for any folder can be determined
          by visiting this folder in the web application
          and copying the ID from the URL. For example,
          for the URL `https://*.app.box.com/folder/123`
          the `folder_id` is `123`.

          The root folder of a Box account is
          always represented by the ID `0`.
        in: path
        name: folder_id
        required: true
        schema:
          type: string
      - description: |-
          A comma-separated list of attributes to include in the
          response. This can be used to request fields that are
          not normally returned in a standard response.

          Be aware that specifying this parameter will have the
          effect that none of the standard fields are returned in
          the response unless explicitly specified, instead only
          fields for the mini representation are returned, additional
          to the fields requested.

          Additionally this field can be used to query any metadata
          applied to the file by specifying the `metadata` field as well
          as the scope and key of the template to retrieve, for example
          `?field=metadata.enterprise_12345.contractTemplate`.
        explode: false
        in: query
        name: fields
        schema:
          items:
            type: string
          type: array
        style: form
      - description: |-
          Specifies whether to use marker-based pagination instead of
          offset-based pagination. Only one pagination method can
          be used at a time.

          By setting this value to true, the API will return a `marker` field
          that can be passed as a parameter to this endpoint to get the next
          page of the response.
        in: query
        name: usemarker
        schema:
          type: boolean
      - description: |-
          Defines the position marker at which to begin returning results. This is
          used when paginating using marker-based pagination.

          This requires `usemarker` to be set to `true`.
        in: query
        name: marker
        schema:
          type: string
      - description: |-
          The offset of the item at which to begin the response.

          Queries with offset parameter value
          exceeding 10000 will be rejected
          with a 400 response.
        in: query
        name: offset
        schema:
          default: 0
          format: int64
          type: integer
      - description: The maximum number of items to return per page.
        in: query
        name: limit
        schema:
          format: int64
          maximum: 1000
          type: integer
      - description: |-
          The URL, and optional password, for the shared link of this item.

          This header can be used to access items that have not been
          explicitly shared with a user.

          Use the format `shared_link=[link]` or if a password is required then
          use `shared_link=[link]&shared_link_password=[password]`.

          This header can be used on the file or folder shared, as well as on any files
          or folders nested within the item.
        in: header
        name: boxapi
        schema:
          type: string
      - description: |-
          Defines the **second** attribute by which items
          are sorted.

          Items are always sorted by their `type` first, with
          folders listed before files, and files listed
          before web links.

          This parameter is not supported for marker-based pagination
          on the root folder (the folder with an ID of `0`).
        in: query
        name: sort
        schema:
          enum:
          - id
          - name
          - date
          - size
          type: string
      - description: The direction to sort results in. This can be either in alphabetical
          ascending (`ASC`) or descending (`DESC`) order.
        in: query
        name: direction
        schema:
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content: {}
          description: "Returns a collection of files, folders, and web links contained\
            \ in a folder."
        "403":
          content: {}
          description: Returned when the access token provided in the `Authorization`
            header is not recognized or not provided.
        "404":
          content: {}
          description: "Returned if the folder is not found, or the user does not\
            \ have access to the folder."
        "405":
          content: {}
          description: Returned if the `folder_id` is not in a recognized format.
        default:
          content: {}
          description: An unexpected client error.
      summary: List items in folder
      tags:
      - Folders
      x-accepts: application/json
      x-tags:
      - tag: Folders
components:
  schemas: {}
x-original-swagger-version: "2.0"